remove return code from repository url and modify README
