+++ README.md
@@ -1,2 +1,23 @@
 # node-template
+
 A template to begin my node project.
+
+1. Clone this repository and you can configure by yourself.
+
+2. Build
+
+	```sh
+	pnpm build
+	```
+
+3. link
+
+	```sh
+	npm link
+	```
+
+4. Run (in any directory)
+
+	```sh
+	node-init
+	```
